+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchanting/EnchantingFeature.java
@@ -203,6 +203,7 @@ public class EnchantingFeature extends JsonFeature {
             new EnchantmentData("iguanatweaksexpanded:air_born").costPerLevel(3),
             new EnchantmentData("iguanatweaksexpanded:recovery").costPerLevel(2),
             new EnchantmentData("iguanatweaksexpanded:vindication").costPerLevel(2),
+            new EnchantmentData("iguanatweaksexpanded:invincible").costPerLevel(4),
 
             new EnchantmentData("minecraft:swift_sneak").costPerLevel(2),
             new EnchantmentData("iguanatweaksexpanded:magnetic").costPerLevel(2),